refactor: Step1, 2에서 공통 사용되는 header 컴포넌트로 분리 (#31)
