+++ src/pages/myPage/ui/customize/step1/components/index.ts
@@ -1,1 +1,1 @@
-export { default as StepGuide } from './StepGuide'
+export { default as StepHeader } from './StepHeader'
